+++ src/components/Choices.jsx
@@ -1,0 +1,3 @@
+export function Choices({ choice, choiceMessage }) {
+  if (choice != null) return <p className="text-xl mt-4">{choiceMessage}</p>;
+}
