+++ src/context/GameContext.jsx
@@ -1,0 +1,73 @@
+import { createContext, useState, useEffect } from "react";
+export const GameContext = createContext();
+import { getResult } from "../helpers/getResult";
+
+import { options } from "../data/options";
+
+export function GameContextProvider(props) {
+  const [userChoice, setUserChoice] = useState(null);
+  const [computerChoice, setComputerChoice] = useState(null);
+  const [userMessage, setUserMessage] = useState(null);
+  const [computerMessage, setComputerMessage] = useState(null);
+  const [result, setResult] = useState(null);
+  const [disabled, setDisabled] = useState(false);
+
+  useEffect(() => {
+    if (userChoice != null) {
+      setUserMessage(
+        `Has elegido ${options[userChoice]?.emoji} - ${options[userChoice]?.name}`
+      );
+    }
+  }, [userChoice]);
+
+  useEffect(() => {
+    if (computerChoice != null) {
+      setComputerMessage(
+        `El ordenador elige ${options[computerChoice]?.emoji} - ${options[computerChoice]?.name}`
+      );
+    }
+  }, [computerChoice]);
+
+  const handlePlay = (choice) => {
+    setUserChoice(choice);
+    setDisabled(true);
+    const randomChoice = Math.floor(Math.random() * 5);
+
+    setTimeout(() => {
+      setComputerChoice(randomChoice);
+    }, 1500);
+
+    setTimeout(() => {
+      setResult(getResult(choice, randomChoice));
+    }, 3000);
+
+    clearTimeout();
+  };
+
+  const reset = () => {
+    setUserChoice(null);
+    setComputerChoice(null);
+    setUserMessage(null);
+    setComputerMessage(null);
+    setResult(null);
+    setDisabled(false);
+  };
+
+  return (
+    <GameContext.Provider
+      value={{
+        userChoice,
+        computerChoice,
+        userMessage,
+        computerMessage,
+        result,
+        disabled,
+        handlePlay,
+        reset,
+        options,
+      }}
+    >
+      {props.children}
+    </GameContext.Provider>
+  );
+}
