+++ src/data/options.jsx
@@ -1,0 +1,7 @@
+export const options = [
+  { id: 0, name: "Piedra", emoji: "🤜", beats: [2, 3] },
+  { id: 1, name: "Papel", emoji: "📄", beats: [0] },
+  { id: 2, name: "Tijera", emoji: "✂️", beats: [1, 3] },
+  { id: 3, name: "Lagarto", emoji: "🦎", beats: [1] },
+  { id: 4, name: "Spock", emoji: "🖖", beats: [3, 0] },
+];
