+++ src/main.jsx
@@ -1,11 +1,13 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
-import Game from "./Game";
-
-import "./index.css";
+import { Game } from "./Game";
+import "./styles.css";
+import { GameContextProvider } from "./context/GameContext";
 
 ReactDOM.createRoot(document.getElementById("root")).render(
   <React.StrictMode>
-    <Game />
+    <GameContextProvider>
+      <Game />
+    </GameContextProvider>
   </React.StrictMode>
 );
